+++ frontend/src/components/index.js
@@ -1,0 +1,5 @@
+import { DataPopup } from "./DataPopup";
+import { PlaceMarker } from "./PlaceMarker";
+import { SearchBar } from "./SearchBar";
+
+export { SearchBar, DataPopup, PlaceMarker }
