+++ frontend/src/pages/Signup.jsx
@@ -1,20 +1,20 @@
 import React, { useState } from 'react'
 import { Link, useNavigate } from 'react-router-dom'
-import MapImg from '../img/map-image.png'
+import MapImg from '/img/map-image.png'
 import axios from 'axios';
 
 const username = 'admin';
 const password = 'admin';
 const token = btoa(`${username}:${password}`);
 
-function Login() {
+function Signup() {
   const navigate = useNavigate();
 
-  const[formData, setFormData] = useState({
-    mail: '',
-    pwd: ''
-  }
-);
+  const [formData, setFormData] = useState({
+    email: '',
+    username: '',
+    password: ''
+  });
 
   const handleChange = (e) => {
     const { name, value } = e.target;
@@ -24,29 +24,31 @@ function Login() {
     });
   }
 
-  const handleLogin = (e) => {
+  const handleSubmit = async (e) => {
     e.preventDefault();
 
-    axios.get(`http://localhost:8080/api/u/auth/login?mail=${formData.mail}&pwd=${formData.pwd}`, 
-            { 
+    axios.post("http://localhost:8080/api/u/auth/register",
+            formData,
+            {
               headers: {
-              'Authorization': `Basic ${token}`
-            }}
-          )
-          .then((response) => {
-            if(response.status == 200) {
-              navigate('/');
-            }
-            console.log(response);
-          })
-          .catch((error) => {``
-            console.log(error);
-          });
-    console.log(formData);    
+                'Authorization': `Basic ${token}`
+              }
+            })
+            .then((response) => {
+              console.log(response);
+              if (response.status === 201) {
+                navigate('/');
+              }
+            })
+            .catch((error) => {
+              console.error(error);
+        });
+
+    console.log(formData);
   }
 
   return (
-    <form onSubmit={handleLogin} className="w-full h-full flex flex-col md:flex-row items-start">
+    <form className="w-full h-full flex flex-col md:flex-row items-start" onSubmit={handleSubmit}>
       <div className="relative h-1/2 md:h-full md:flex-1 flex items-center justify-center flex-col">
         <div className="absolute flex flex-col text-center md:text-left logo-text">
           <h1 className="text-2xl md:text-3xl text-[#060606] font-bold my-2">Wherever You Go,</h1>
@@ -59,15 +61,23 @@ function Login() {
 
         <div className="w-full flex flex-col max-w-[500px]">
           <div className="w-full flex flex-col mb-2">
-            <h3 className="text-2xl md:text-3xl font-semibold mb-2 text-black">Login</h3>
-            <p className="text-base mb-2 text-[#64BE8F]">Welcome back!</p>
+            <h3 className="text-2xl md:text-3xl font-semibold mb-2 text-black">Signup</h3>
+            <p className="text-base mb-2 text-[#64BE8F]">Create Your Account</p>
           </div>
 
           <div className="w-full flex flex-col">
             <input
-              type="text"
+              type="email"
               placeholder="Email"
-              name="mail"
+              name='email'
+              value={formData.email}
+              onChange={handleChange}
+              className="w-full text-black py-2 my-2 bg-transparent border-b border-black outline-none focus:outline-none"
+            />
+            <input
+              type="text"
+              placeholder="Username"
+              name='username'
               value={formData.username}
               onChange={handleChange}
               className="w-full text-black py-2 my-2 bg-transparent border-b border-black outline-none focus:outline-none"
@@ -75,8 +85,8 @@ function Login() {
             <input
               type="password"
               placeholder="Password"
-              name="pwd"
-              value={formData.pwd}
+              name='password'
+              value={formData.password}
               onChange={handleChange}
               className="w-full text-black py-2 my-2 bg-transparent border-b border-black outline-none focus:outline-none"
             />
@@ -86,13 +96,11 @@ function Login() {
               <input type="checkbox" className="w-4 h-4 mr-2" />
               <p className="text-sm text-black">Remember me</p>
             </div>
-
-            <p className="text-sm font-medium whitespace-nowrap cursor-pointer underline underline-offset-2 text-black">Forget Password</p>
           </div>
           <div className="w-full flex flex-col my-4">
-          <button type="submit" className="w-full text-white my-2 font-semibold bg-[#5DB487] rounded-md p-4 text-center flex items-center justify-center cursor-pointer">
-            Login
-          </button>
+          <button type='submit' className="w-full text-white my-2 font-semibold bg-[#5DB487] rounded-md p-4 text-center flex items-center justify-center cursor-pointer">
+              Signup
+            </button>
           </div>
 
           <div className="w-full flex items-center justify-center relative py-2">
@@ -100,13 +108,11 @@ function Login() {
             <p className="text-lg absolute text-black/80 bg-[#f5f5f5]">or</p>
           </div>
         </div>
-        
 
         <div className="w-full flex items-center justify-center">
-          <p className="text-sm font-normal text-[#060606]">
-            Don't have an account?
-            <Link to="/signup" className="font-semibold underline underline-offset-2 cursor-pointer text-[#5DB487]">
-              Register
+          <p className="text-sm font-normal text-[#060606]">Already have an account?
+            <Link to="/login" className="font-semibold underline underline-offset-2 cursor-pointer text-[#5DB487]">
+              Login
             </Link>
           </p>
         </div>
@@ -115,4 +121,4 @@ function Login() {
   );
 }
 
-export { Login }
+export { Signup }
