+++ frontend/src/pages/admin/ViewPlaces.jsx
