+++ frontend/src/pages/index.jsx
@@ -2,6 +2,4 @@ export { Login } from './Login.jsx';
 export { Signup } from './Signup.jsx';
 export { Home } from './Home.jsx';
 export { MapComponent } from './MapComponent.jsx';
-export { SearchBar } from './SearchBar.jsx';
-export { DataPopup } from './DataPopup.jsx';
 export { AdminDashboard } from './AdminDashboard.jsx';
