+++ frontend/src/utils/MarkerIcons.jsx
@@ -2,7 +2,7 @@ import L from 'leaflet';
 import markerIcon2x from 'leaflet/dist/images/marker-icon-2x.png';
 import markerIcon from 'leaflet/dist/images/marker-icon.png';
 import markerShadow from 'leaflet/dist/images/marker-shadow.png';
-import userIco from '../img/user.png';
+import userIco from '/img/user.png';
 
 const smallIcon = L.icon({
   iconUrl: markerIcon,
